web app/Smart Ro/Smart-Ro/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.13.0
django-heroku==0.3.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
twilio==7.9.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
